Задание 3 lengTH

+++ Program.cs
@@ -1,5 +1,5 @@
-﻿int [] array = { 4, -3, 7 };
-int ind=0, size = array.Length;
+﻿long [] array = { 4, -3, 7 };
+long ind=0, size = array.Length;
 while (ind<size) {
 if (array[ind] % 2) {
 Console.Write(array[ind]);
